moved put endpoint tests to endpoints/put_test.go
